// weather-app/src/app/interfaces/displayed-weather-data.interface.ts
export interface DisplayedWeatherData {
    app_temp: number;
    clouds: number;
    wind_spd: number;
    pod: string;
    pres: number;
    rh: number;
    sunrise: string;
    sunset: string;
    temp: number;
    wind_cdir_full: string;
    wind_dir: number;
    weather: string;
    city_name: string
  }